[JAMES][ADD][CONTACTS/STATUS]

+++ MoneyGer.Server/Controllers/CompanyController.cs
@@ -2,6 +2,7 @@ using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.DotNet.Scaffolding.Shared.Messaging;
 using Microsoft.EntityFrameworkCore;
 using MoneyGer.Server.Context;
 using MoneyGer.Server.Dtos;
@@ -9,6 +10,7 @@ using MoneyGer.Server.Models;
 
 namespace MoneyGer.Server.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class CompanyController : ControllerBase
@@ -32,18 +34,19 @@ namespace MoneyGer.Server.Controllers
                 return BadRequest("Company name is required");
             }
             
+            var owner = User.FindFirstValue(ClaimTypes.NameIdentifier);
             var company = new Company
             {
                 Id = Guid.NewGuid().ToString(),
                 Name = createCompanyDto.Name,
                 Location = createCompanyDto.Location,
-                Owner = createCompanyDto.Owner!
+                Owner = owner!
             };
 
             _context.companies.Add(company);
             await _context.SaveChangesAsync();
 
-            var user = await _userManager.FindByIdAsync(createCompanyDto.Owner!);
+            var user = await _userManager.FindByIdAsync(owner!);
 
             if(user != null){
 
@@ -82,8 +85,22 @@ namespace MoneyGer.Server.Controllers
             {
                 return NotFound("Company not Found");
             }
+            var user = await _userManager.FindByIdAsync(company.Owner);
+
+            if (user == null)
+            {
+                return NotFound("User not Found");
+            }
+
+            user.Company ="N/A";
 
             _context.companies.Remove(company);
+            
+           var result = await _userManager.UpdateAsync(user); //Updates 'Company' column for user info
+
+                if (!result.Succeeded)
+                    return BadRequest(result);
+
             await _context.SaveChangesAsync();
 
             return Ok(new { message = "Company deleted" });
@@ -114,34 +131,66 @@ namespace MoneyGer.Server.Controllers
             }
 
             // Check if the user already has a role assigned in the specified company
-            var existingUserRole = await _context.UserCompanyRole.FirstOrDefaultAsync(ucr => ucr.UserId == user.Id && ucr.CompanyId == company.Id);
+            var existingUserRole = await _context.UserCompanyRole.FirstOrDefaultAsync(ucr => ucr.UserId == user.Id);
 
             if (existingUserRole != null)
             {
-                // Update the role for the user in the specified company
-                existingUserRole.RoleId = role.Id;
+                 _context.UserCompanyRole.Remove(existingUserRole);
+                await _context.SaveChangesAsync();
             }
-            else
+            
+            // Create a new entry for the user in the specified company
+            var newUserRole = new UserCompanyRole
             {
-                // Create a new entry for the user in the specified company
-                var newUserRole = new UserCompanyRole
-                {
-                    UserId = user.Id,
-                    RoleId = role.Id,
-                    CompanyId = company.Id
-                };
+                UserId = user.Id,
+                RoleId = role.Id,
+                CompanyId = company.Id
+            };
 
-                _context.UserCompanyRole.Add(newUserRole);
-                
+            try
+            {
+                _context.UserCompanyRole.Add(newUserRole); //Add newUserRole to the database
+
                 user.Company = company.Name;
-                var result = await _userManager.UpdateAsync(user);
+                var result = await _userManager.UpdateAsync(user); //Updates 'Company' column for user info
 
-                if(!result.Succeeded)
+                if (!result.Succeeded)
                     return BadRequest(result);
+
+                await _context.SaveChangesAsync();
             }
-            await _context.SaveChangesAsync();
+
+            catch (Exception e)
+            {
+                Console.WriteLine(e.StackTrace);
+            }
 
             return Ok(new { message = "Company assignment successful!" });
+
+        }
+
+        [HttpDelete("DeleteUser")]
+        public async Task<IActionResult> DeleteUserCompany([FromBody] CompanyAssignDto companyAssignDto)
+        {
+            var companyRole = await _context.UserCompanyRole.FindAsync(companyAssignDto.UserId, companyAssignDto.RoleId, companyAssignDto.CompanyId);
+
+            if (companyRole == null)
+            {
+                return NotFound(new { message = "User not found" });
+            }
+
+            _context.UserCompanyRole.Remove(companyRole);
+                    
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (Exception e)
+            {
+                return BadRequest(new { message = e.Message });
+            }
+
+            return Ok(new { message = "User deleted" });
         }
     }
 }
